cull widgets outside of scissor areas
