+++ src/main/java/com/cleanroommc/modularui/api/widget/IWidget.java
@@ -2,6 +2,7 @@ package com.cleanroommc.modularui.api.widget;
 
 import com.cleanroommc.modularui.api.ITheme;
 import com.cleanroommc.modularui.api.layout.ILayoutWidget;
+import com.cleanroommc.modularui.drawable.Scissor;
 import com.cleanroommc.modularui.screen.ModularPanel;
 import com.cleanroommc.modularui.screen.viewport.GuiContext;
 import com.cleanroommc.modularui.theme.WidgetTheme;
@@ -109,8 +110,9 @@ public interface IWidget extends IGuiElement {
 
     void setEnabled(boolean enabled);
 
-    // TODO: Really needed?
-    boolean canBeSeen();
+    default boolean canBeSeen() {
+        return Scissor.isInsideScissorArea(getArea(), getContext());
+    }
 
     default boolean canHover() {
         return true;
